tests2/test_benchmark.py
# import unittest
# from src.benchmarking.database import MetricsDatabase, BenchmarkMetrics, BenchmarkSessions

# class TestMetricsDatabase(unittest.TestCase):
#     @classmethod
#     def setUpClass(cls):
#         cls.db = MetricsDatabase(":memory:")
    
#     @classmethod
#     def tearDownClass(cls):
#         cls.db.conn.close()

#     def test_create_benchmark_metrics(self):
#         mock_data = BenchmarkMetrics(puzzle_size="5x5", algorithm="Classical", steps=50, time=0.5)
#         self.db.create_benchmark_metrics(mock_data)
#         record = self.db.get_benchmark_metrics(mock_data.id)
#         self.assertEqual(record.puzzle_size, "5x5")
#         self.assertEqual(record.algorithm, "Classical")
#         self.assertEqual(record.steps, 50)
#         self.assertEqual(record.time, 0.5)

#     def test_update_benchmark_metrics(self):
#         mock_data = BenchmarkMetrics(puzzle_size="5x5", algorithm="Classical", steps=50, time=0.5)
#         self.db.create_benchmark_metrics(mock_data)
#         mock_data.steps = 60
#         self.db.update_benchmark_metrics(mock_data)
#         record = self.db.get_benchmark_metrics(mock_data.id)
#         self.assertEqual(record.steps, 60)

#     def test_delete_benchmark_metrics(self):
#         mock_data = BenchmarkMetrics(puzzle_size="5x5", algorithm="Classical", steps=50, time=0.5)
#         self.db.create_benchmark_metrics(mock_data)
#         self.db.delete_benchmark_metrics(mock_data.id)
#         record = self.db.get_benchmark_metrics(mock_data.id)
#         self.assertIsNone(record)

#     def test_create_benchmark_session(self):
#         mock_data = BenchmarkSessions(session_id=1, num_boards=500, puzzle_size="5x5", algorithm="Classical")
#         self.db.create_benchmark_session(mock_data)
#         record = self.db.get_benchmark_session(mock_data.session_id)
#         self.assertEqual(record.num_boards, 500)

#     # Create additional tests for update, delete and retrieval for BenchmarkSessions similar to the tests for BenchmarkMetrics

# if __name__ == "__main__":
#     unittest.main()
from src.benchmarking.database import MetricsDatabase, BenchmarkMetrics, BenchmarkSession, BenchmarkMetrics
import sqlite3
import unittest
from src.benchmarking.benchmark import Benchmark, TimeBenchmark
from unittest.mock import Mock, patch
import time

class TestMetricsDatabase(unittest.TestCase):
    @classmethod
    def setUpClass(cls):
        cls.db = MetricsDatabase(":memory:")
        
    def test_create_benchmark_metrics(self):
        mock_data = BenchmarkMetrics(1, "5x5", "algorithm1", 50, 2.5)
        self.db.create_benchmark_metrics(mock_data)

        with mock.patch.object(self.db, "get_benchmark_metrics", return_value=mock_data):
            record = self.db.get_benchmark_metrics(1)
            self.assertEqual(record.puzzle_size, "5x5")
        
    def test_update_benchmark_metrics(self):
        mock_data = BenchmarkMetrics(1, "5x5", "algorithm1", 60, 2.7)
        self.db.update_benchmark_metrics(mock_data)
        
        with mock.patch.object(self.db, "get_benchmark_metrics", return_value=mock_data):
            record = self.db.get_benchmark_metrics(1)
            self.assertEqual(record.steps, 60)
        
    def test_delete_benchmark_metrics(self):
        mock_data = BenchmarkMetrics(1, "5x5", "algorithm1", 50, 2.5)
        try:
            self.db.create_benchmark_metrics(mock_data)
        except sqlite3.IntegrityError:
            self.db.delete_benchmark_metrics(mock_data.id)
            self.db.create_benchmark_metrics(mock_data)
        
        self.db.delete_benchmark_metrics(mock_data.id)
        
        with mock.patch.object(self.db, "get_benchmark_metrics", return_value=None):
            record = self.db.get_benchmark_metrics(1)
            self.assertIsNone(record)

class TestTimeBenchmark(unittest.TestCase):
    def setUp(self):
        self.mock_puzzle = Mock()
        self.mock_algorithm = Mock()
        self.mock_algorithm.name = 'mock_algorithm'  
        self.mock_algorithm.run.return_value = (200, 0.2)  # 200 steps, 0.2 seconds
        self.time_benchmark = TimeBenchmark(self.mock_puzzle, self.mock_algorithm)

    # @patch('time.perf_counter')  # TODO: Implement time mocking
    def test_run_method(self): 
        # mock_time.side_effect = [0, 0.2]  
        metrics_data = self.time_benchmark.run()
        self.assertIsInstance(metrics_data, BenchmarkMetrics)
        self.assertEqual(metrics_data.puzzle_size, self.mock_puzzle.size)
        self.assertEqual(metrics_data.algorithm, self.mock_algorithm.name)
        self.assertEqual(metrics_data.steps, 200)
        # self.assertEqual(metrics_data.time, 0.2) # TODO: check once time is mocked

class TestMetricsDatabase(unittest.TestCase):
    def setUp(self):
        self.db = MetricsDatabase(":memory:")

    @patch('src.benchmarking.database.MetricsDatabase.get_benchmark_metrics')
    def test_create_benchmark_metrics(self, mock_get_benchmark_metrics):
        mock_data = BenchmarkMetrics(1, "5x5", "Quantum Algorithm", 50, 0.5)
        mock_get_benchmark_metrics.return_value = mock_data
        self.db.create_benchmark_metrics(mock_data)
        record = self.db.get_benchmark_metrics(mock_data.id)
        self.assertEqual(record, mock_data)

    # Repeat similar steps for test_delete_benchmark_metrics and test_update_benchmark_metrics using patch.
